+++ utils/tools.go
@@ -1,7 +1,6 @@
 package utils
 
 import (
-	"crypto/md5"
 	crand "crypto/rand"
 	"encoding/json"
 	"fmt"
@@ -9,11 +8,6 @@ import (
 	"runtime/debug"
 	"strings"
 )
-
-func Md5(str string) string {
-	t := md5.Sum(([]byte)(str))
-	return fmt.Sprintf("%x", t)
-}
 
 func CopyStructByJson(src, dst interface{}) {
 	str, _ := json.Marshal(src)